refactor: [step1피드백반영] CreateClause 삭제
